fix(logging): remove custom logger name/hierarchy
