+++ plantseg/loggers.py
@@ -1,7 +1,7 @@
 import logging
 import sys
 
-LOGGING_FORMAT = '%(handler_name)s [%(threadName)s] %(asctime)s %(name)s - %(message)s'
+LOGGING_FORMAT = "%(handler_name)s [%(threadName)s] %(asctime)s %(name)s - %(message)s"
 formatter_viewer_napari = logging.Formatter(LOGGING_FORMAT)  # Napari prepends '%(levelname)s: '.upper()
 formatter_default = logging.Formatter('%(levelname)s: ' + LOGGING_FORMAT)
 
@@ -20,9 +20,3 @@ class PlantSegHandler(logging.StreamHandler):
 stream_handler = PlantSegHandler(sys.stdout)
 stream_handler.set_name("P")
 stream_handler.setFormatter(formatter_default)
-
-logger_root = logging.getLogger("PlantSeg")
-logger_root.setLevel(logging.INFO)
-logger_root.addHandler(stream_handler)
-
-logger_logger = logging.getLogger("PlantSeg.Logger")
