+++ plantseg/viewer_napari/widgets/predictions.py
@@ -14,7 +14,7 @@ from napari.types import LayerDataTuple
 from plantseg.models.zoo import model_zoo
 from plantseg.plantseg_image import PlantSegImage
 from plantseg.tasks.predictions_tasks import unet_predictions_task
-from plantseg.viewer_napari.logging import napari_formatted_logging
+from plantseg.viewer_napari import napari_formatted_logging
 from plantseg.viewer_napari.widgets.segmentation import widget_agglomeration, widget_dt_ws, widget_lifted_multicut
 from plantseg.viewer_napari.widgets.utils import schedule_task
