+++ plantseg/viewer_napari/widgets/proofreading.py
@@ -11,7 +11,7 @@ from napari.types import LayerDataTuple
 from plantseg.functionals.proofreading.split_merge_tools import split_merge_from_seeds
 from plantseg.functionals.proofreading.utils import get_bboxes
 from plantseg.plantseg_image import ImageProperties, PlantSegImage, SemanticType
-from plantseg.viewer_napari.logging import napari_formatted_logging
+from plantseg.viewer_napari import napari_formatted_logging
 
 DEFAULT_KEY_BINDING_PROOFREAD = 'n'
 DEFAULT_KEY_BINDING_CLEAN = 'b'
